+++ src/main/java/app/domain/menu/internal/InternalMenuController.java
@@ -6,6 +6,8 @@ import app.domain.menu.status.StoreMenuSuccessStatus;
 import app.global.apiPayload.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,9 +17,11 @@ import java.util.UUID;
 @RequestMapping("/internal/menus")
 @RequiredArgsConstructor
 @Tag(name = "내부 API", description = "내부 API")
+@Slf4j
 public class InternalMenuController {
 
 	private final InternalMenuService internalMenuService;
+	private final StockRetryService stockRetryService;
 
 	@PostMapping("/batch")
 	public ApiResponse<List<MenuInfoResponse>> getMenuInfoList(@RequestBody List<UUID> menuIds) {
@@ -26,7 +30,8 @@ public class InternalMenuController {
 
 	@PostMapping("/stocks/decrease")
 	public ApiResponse<Boolean> decreaseStock(@RequestBody List<StockRequest> requests) {
-		boolean result = internalMenuService.decreaseStock(requests);
+		log.info("InternalMenuService class: {}", internalMenuService.getClass().getName());
+		boolean result = stockRetryService.decreaseStock(requests);
 		return ApiResponse.onSuccess(StoreMenuSuccessStatus.STOCK_DECREASE_SUCCESS, result);
 	}
 }
