+++ src/main/java/app/domain/menu/model/repository/StockRepository.java
@@ -1,7 +1,10 @@
 package app.domain.menu.model.repository;
 
 import app.domain.menu.model.entity.Stock;
+import jakarta.persistence.LockModeType;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Lock;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -11,7 +14,9 @@ import java.util.UUID;
 @Repository
 public interface StockRepository extends JpaRepository<Stock, UUID> {
 
+    @Lock(LockModeType.OPTIMISTIC)
     List<Stock> findByMenuMenuIdIn(List<UUID> menuIds);
 
+    @Lock(LockModeType.OPTIMISTIC)
     Optional<Stock> findByMenu_MenuId(UUID menuId);
 }
